+++ src/realm/object-store/c_api/realm.cpp
@@ -19,7 +19,7 @@ RLM_API bool realm_get_version_id(const realm_t* realm, bool* out_found, realm_v
     });
 }
 
-RLM_API bool realm_get_num_versions(const realm_t* realm, size_t* out_count)
+RLM_API bool realm_get_num_versions(const realm_t* realm, uint64_t* out_count)
 {
     return wrap_err([&]() {
         *out_count = (*realm)->get_number_of_versions();
